Add vuex-persistedstate
