+++ movie_vue/src/views/accounts/SignupView.vue
@@ -31,7 +31,7 @@ export default {
         }
     },
     methods: {
-        ...mapActions(['signup'])
+        ...mapActions('CommunityModule',['signup'])
     }
 }
 </script>
