+++ movie_vue/src/views/movies/MovieDeleteView.vue
@@ -10,7 +10,7 @@ import { mapActions } from 'vuex'
 export default {
     name: 'MovieDeleteView',
     methods: {
-        ...mapActions(['deleteMovie'])
+        ...mapActions('CommunityModule',['deleteMovie'])
     },
     created() {
         this.deleteMovie(this.$route.params.id)
