+++ movie_vue/src/views/movies/MovieListView.vue
@@ -4,7 +4,6 @@
       <div class="movie-list-item row">
         <div class="col-3" v-for="movie in movies" :key="movie.id">
             <div class="card">
-                <p>{{ movie.poster_path }}</p>
                 <img :src="'http://image.tmdb.org/t/p/w185/'+ movie.poster_path" class="card-img-top" alt="영화 포스터">
                 <div class="card-body">
                     <h5 class="card-title">{{ movie.title }} <span class="badge badge-info">{{ movie.vote_average }}</span></h5>
@@ -23,15 +22,15 @@ import { mapActions,mapState } from 'vuex'
 export default {
     name:'MovieListView',
     computed: {
-        ...mapState(['movies'])
+        ...mapState('CommunityModule',['movies'])
     },
     methods: {
-        ...mapActions(['fetchMovies']),
+        ...mapActions('CommunityModule',['fetchMovies']),
         movieDetail(movieId) {
             this.$router.push({ name: 'MovieDetail', params: { id: movieId}})
         },
     },
-    created() {
+    mounted() {
         this.fetchMovies()
     }
 }
